renaming email module
