+++ rdr_service/genomic/genomic_job_controller.py
@@ -56,7 +56,7 @@ from rdr_service.dao.genomics_dao import (
 from rdr_service.resource.generators.genomics import genomic_job_run_update, genomic_file_processed_update, \
     genomic_manifest_file_update, genomic_manifest_feedback_update, genomic_gc_validation_metrics_batch_update, \
     genomic_set_member_batch_update
-from rdr_service.services.email import Email, EmailService
+from rdr_service.services.email_service import Email, EmailService
 from rdr_service.services.slack_utils import SlackMessageHandler
