+++ rdr_service/services/email_service.py
